抽号的一些小修改 增加了实体类

+++ src/main/java/com/iqes/repository/queue/QueueManagerDao.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.ArrayList;
 import java.util.List;
 
-public interface QueueManagerDao extends CrudRepository<QueueInfo, Integer> {
+public interface QueueManagerDao extends CrudRepository<QueueInfo, Long> {
 
     //获取同桌类型的排队顾客
     @Query("select q from QueueInfo q where q.tableType=(select t.tableType from TableNumber t where t.name=?1) order by q.id")
